+++ src/poll.js
